Frontend/requirements.txt
click==6.7
Flask==0.12.2
Flask-HTTPAuth==3.2.3
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
passlib==1.7.1
SQLAlchemy==1.1.11
Werkzeug==0.12.2
pyopenssl
pyyaml
